# api/docs/swagger.yaml
definitions:
...
  user.UserDetailResp:
    properties:
      name:
        type: string
      uuid:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/user/detail:
    get:
      parameters:
      - in: query
        name: uuid
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserDetailResp'
      summary: 用户详情
      tags:
      - user
swagger: "2.0"